backend/requirements.txt
feedparser
python-dateutil
openai
requests